Use autosummary to generate API docs
Fixes #50

Use autosummary to generate the full API docs for each module. rst pages
for each module are created automatically. Each function/class has it's
own page to make it easier to find what you want.

This only works with imported members (in the `__init__.py` files)
because I separated the build in 2: 1) generate the rst stubs with
sphinx-autogen and the `--imported-members` flag 2) build the HTML with
sphinx-build. This is a recent change in sphinx and using
`autosummary_generate = True` in `conf.py` won't work
(sphinx-doc/sphinx#3235).

The stub pages are created in the `generated` folder. Added this folder
to `make clean` and included some new rules in the Makefile:
* `api`: runs sphinx-autogen
* `serve`: serve the `_build/html` folder on `localhost:8015` (useful
 for previewing)

It would be good to add a 1-sentence summary of each package in the
docstrings to populate the table of modules that autosummary makes.